move function to single_star to stop cyclical imports
