+++ posydon/binary_evol/DT/step_merged.py
@@ -11,8 +11,8 @@ __authors__ = [
 import numpy as np
 
 from posydon.utils.data_download import PATH_TO_POSYDON_DATA
-from posydon.binary_evol.singlestar import STARPROPERTIES
-from posydon.utils.common_functions import check_state_of_star, convert_star_to_massless_remnant
+from posydon.binary_evol.singlestar import STARPROPERTIES, convert_star_to_massless_remnant
+from posydon.utils.common_functions import check_state_of_star
 from posydon.binary_evol.DT.step_isolated import IsolatedStep
 
 import warnings
